chore: Move login class in sub package
