+++ src/commonTest/kotlin/io/github/hansanto/kault/auth/approle/VaultAuthAppRoleTest.kt
@@ -9,7 +9,7 @@ import io.github.hansanto.kault.auth.approle.response.AppRoleLookUpSecretIdRespo
 import io.github.hansanto.kault.auth.approle.response.AppRoleReadRoleIdResponse
 import io.github.hansanto.kault.auth.approle.response.AppRoleReadRoleResponse
 import io.github.hansanto.kault.auth.approle.response.AppRoleWriteSecretIdResponse
-import io.github.hansanto.kault.auth.common.LoginResponse
+import io.github.hansanto.kault.auth.common.response.LoginResponse
 import io.github.hansanto.kault.exception.VaultAPIException
 import io.github.hansanto.kault.system.auth.enable
 import io.github.hansanto.kault.util.DEFAULT_ROLE_NAME
